Add method to optimize for relative errors.
-Select half of the eigenvectors as before, and for the other half do:
--Compute the ratio between the current reduced diagonal of the
covariance matrix and that of the prior (ie ratio of variances at each
point) in X and select the smallest one (the worse).
--Select the eigenvector that gives the highest contribution to that
diagonal element (which is computed from <x|u>s²<u|x> for each
eigenvector u, if x is the worst point). Add this to the selected
eigenvectors and repeat the procedure.

This behaviour can be switched on and off by setting the value of
compress_X.